LAB05_FINAL and LAB06_FINAL
I merged it in one file because there is only minor adjustments for lab06 from lab05. Both Lab05 and Lab06 find the critical path and critical time. The only difference is Lab06 considers the edge weight.

LAB04 is also uploaded with minor adjustments in the Tester class due to some adjustments in the DGraph file.